src/requirements.txt
ghostscript==0.7
matplotlib==3.8.3
numpy==1.26.4
parameterized==0.9.0
Pillow==10.2.0
sv_ttk==2.6.0
